Separa GenericDAO e divide Tempo de Tramitacao em duas versões
