+++ dominio/tutela/queries/tempo_tramitacao.sql
@@ -1,0 +1,2 @@
+SELECT * FROM {schema}.tb_tempo_tramitacao
+WHERE id_orgao = :orgao_id
