+++ dominio/tutela/queries/tempo_tramitacao_integrado.sql
@@ -1,0 +1,2 @@
+SELECT * FROM {schema}.tb_tempo_tramitacao_integrado
+WHERE id_orgao = :orgao_id
